+++ text/truncate.go
@@ -1,0 +1,15 @@
+package text
+
+// Truncate limits a string to the specified number of characters.
+func Truncate(value string, max int) string {
+	if max <= 0 {
+		return ""
+	}
+
+	runes := []rune(value)
+	if len(runes) <= max {
+		return value
+	}
+
+	return string(runes[:max])
+}